beta994

+++ Nuget-AspNetCore.sh
@@ -27,10 +27,33 @@ dotnet pack SimpleAuth.Client/SimpleAuth.Client.csproj -c Release -p:PackageVers
 dotnet pack SimpleAuth.Shared/SimpleAuth.Shared.csproj -c Release -p:PackageVersion=$VERSION -o .
 dotnet pack SimpleAuth.Core/SimpleAuth.Core.csproj -c Release -p:PackageVersion=$VERSION -o .
 
-dotnet nuget push SimpleAuth.Client.AspNetCore.$VERSION.nupkg $NUGET_SERVER -s $NUGET_URL
+dotnet nuget push SimpleAuth.Client.AspNetCore.$VERSION.nupkg -k $NUGET_SERVER -s $NUGET_URL
+if [ $? -ne 0 ]
+then
+    echo 'Failure pushing package SimpleAuth.Client.AspNetCore to nuget server'
+    exit 2
+fi
+
 dotnet nuget push SimpleAuth.Client.$VERSION.nupkg -k $NUGET_SERVER -s $NUGET_URL
+if [ $? -ne 0 ]
+then
+    echo 'Failure pushing package SimpleAuth.Client to nuget server'
+    exit 2
+fi
+
 dotnet nuget push SimpleAuth.Shared.$VERSION.nupkg -k $NUGET_SERVER -s $NUGET_URL
+if [ $? -ne 0 ]
+then
+    echo 'Failure pushing package SimpleAuth.Shared to nuget server'
+    exit 2
+fi
+
 dotnet nuget push SimpleAuth.Core.$VERSION.nupkg -k $NUGET_SERVER -s $NUGET_URL
+if [ $? -ne 0 ]
+then
+    echo 'Failure pushing package SimpleAuth.Core to nuget server'
+    exit 2
+fi
 
 rm -f *.nupkg
